rebase fix 3

+++ src/engine/matrix.cpp
@@ -28,16 +28,20 @@ void cross(float *a, float *b, float *res) {
 }
 
 
-void normalize(float *a) {
-        float l {sqrt(a[0] * a[0] + a[1] * a[1] + a[2] * a[2])};
+void
+normalize(float *a)
+{
+        float l {static_cast<float>(sqrt(a[0] * a[0] + a[1] * a[1] + a[2] * a[2]))};
         a[0] = a[0] / l;
         a[1] = a[1] / l;
         a[2] = a[2] / l;
 }
 
 
-float length(float *v) {
-        float res {sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2])};
+float
+length(float *v)
+{
+        float res {static_cast<float>(sqrt(v[0] * v[0] + v[1] * v[1] + v[2] * v[2]))};
         return res;
 }
